test(pty-daemon): harden fd-release regression tests per self-review
- ptyFdCount now throws when lsof yields no output instead of returning
  0 — previously a runner without lsof made every assertion pass
  vacuously, so a reverted patch would still go green.
- Count Linux pty slaves too (/dev/pts/) so the no-leak assertions
  measure something off-macOS.
- Compute the leak count for failure messages only on the failure path
  (the eagerly-evaluated template ran an extra lsof per assertion).
- Reuse a shared test/helpers/wait-for.ts poller instead of adding a
  third copy of the waitFor helper; foreground-process.test.ts now
  imports it too.
- Shorten the background child to sleep 5 so the suite doesn't leave
  30-second orphans behind.
